Add abstract superclass for the history, today and future controllers
